requirements.txt
Django
djangorestframework
tensorflow
lazypredict
django-cors-headers
djangorestframework-simplejwt